refactor Time into its own file, fix lints
