+++ package.json
@@ -8,7 +8,7 @@
   ],
   "scripts": {
     "preinstall": "cd mamar-wasm-bridge && wasm-pack build -t web && cd ../pm64-typegen && cargo run",
-    "lint": "cargo +nightly fmt --all --check && cd mamar-web && yarn run lint"
+    "lint": "cargo fmt --all --check && cd mamar-web && yarn run lint"
   },
   "multipleStaticFileCopier": [
     {
